refactor: DTO 패키지 컨벤션에 맞게 수정
