+++ module-core/module-client/src/main/kotlin/dto/response/GithubRepositoryResponse.kt
@@ -1,4 +1,4 @@
-package dto
+package dto.response
 
 import com.fasterxml.jackson.annotation.JsonProperty
